requirements.txt
asgiref==3.6.0
Django==3.2.18
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Faker==18.3.2
faker-vehicle==0.2.0
psycopg2==2.9.5
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
